git-svn-id: http://opensvn.wiimm.de/wii/trunk/wiimms-iso-tools@1386 4ca24660-9720-11de-8e73-472ee046f03b

+++ src/lib-ciso.c
@@ -135,7 +135,7 @@ u32 CalcBlockSizeCISO ( u32 * result_n_blocks, off_t file_size )
 	if ( file_size < temp )
 	    file_size = temp;
     }
-    TRACE(" - file_size  := %llx\n",file_size);
+    TRACE(" - file_size  := %llx\n",(u64)file_size);
 
     u64 block_size = opt_chunk_size;
     if ( !block_size || !force_chunk_size )
